% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodies.R
\name{dyak_body_annotations}
\alias{dyak_body_annotations}
\title{Return neurojson body annotations via the Clio interface}
\usage{
dyak_body_annotations(
  ids = NULL,
  query = NULL,
  json = FALSE,
  config = NULL,
  show.extra = c("none", "user", "time", "all"),
  cache = FALSE,
  test = FALSE,
  ...,
  dataset = dyak_default_dataset()
)
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link[malevnc]{manc_ids}}}

\item{query}{A json query string (see examples or documentation) or an R list
with field names as elements.}

\item{json}{Whether to return unparsed JSON rather than an R list (default
\code{FALSE}).}

\item{config}{An optional httr::config (expert use only, must include a
bearer token)}

\item{show.extra}{Extra columns to show with user/timestamp information.}

\item{cache}{Whether to cache the result of this call for 5 minutes.}

\item{test}{Whether to unset the clio-store test server (default
\code{FALSE})}

\item{...}{Additional arguments passed to \code{\link[pbapply:pblapply]{pbapply::pblapply()}}.}

\item{dataset}{The name of the dataset as reported in Clio e.g. \code{yakuba},
\code{yakuba-vnc} etc. \code{yakuba}, \code{yakuba-vnc} and \code{yakubavnc} are treated as
aliases for the main yakuba dataset.}
}
\value{
A data frame with metadata.
}
\description{
Return neurojson body annotations via the Clio interface
}
\details{
In comparison with \code{dyak_neuprint_meta()}, this provides access to
up to the second annotations; it is also presently faster than that call.
Compared with \code{dyak_neuprint_meta()}, it does not produce a stable set of
columns, only returning those that exist for the given query ids.
}
\examples{
\donttest{
dyak_body_annotations("DNa02")
}
}
\seealso{
\code{\link[malevnc:manc_body_annotations]{malevnc::manc_body_annotations()}}
}
