% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{dyak_connection_table}
\alias{dyak_connection_table}
\title{Connectivity query for yakuba neurons}
\usage{
dyak_connection_table(
  ids,
  partners = c("inputs", "outputs"),
  moredetails = c("class", "group", "type", "somaSide", "somaNeuromere", "hemilineage"),
  summary = FALSE,
  threshold = 1L,
  roi = NULL,
  by.roi = FALSE,
  conn = NULL,
  ...,
  dataset = dyak_default_dataset()
)
}
\arguments{
\item{ids}{A set of body ids (see \code{\link[=dyak_ids]{dyak_ids()}} for a range of ways to
specify these).}

\item{partners}{Either inputs or outputs. Redundant with \code{prepost}, but
probably clearer.}

\item{moredetails}{Either a logical (to add all fields when \code{TRUE}) or a
character vector naming additional fields returned by
\code{dyak_neuprint_meta()} that will be added to the results data frame.}

\item{summary}{When \code{TRUE} and more than one query neuron is given,
summarises connectivity grouped by partner.}

\item{threshold}{Only return partners >= to an integer value. Default of 1
returns all partners. This threshold will be applied to the ROI weight when
the \code{roi} argument is specified, otherwise to the whole neuron.}

\item{roi}{a single ROI. Use \code{neuprint_ROIs} to see what is available.}

\item{by.roi}{logical, whether or not to break neurons' connectivity down by
region of interest (ROI)}

\item{conn}{Optional, a \code{neuprint_connection} object. Defaults to
\code{\link[=dyak_neuprint]{dyak_neuprint()}} to ensure that query is against the yakuba dataset.}

\item{...}{additional arguments passed to
\code{\link{neuprint_connection_table}}}

\item{dataset}{The name of the dataset as reported in Clio e.g. \code{yakuba},
\code{yakuba-vnc} etc. \code{yakuba}, \code{yakuba-vnc} and \code{yakubavnc} are treated as
aliases for the main yakuba dataset.}
}
\value{
A data frame.
}
\description{
Connectivity query for yakuba neurons
}
\examples{
\donttest{
dyak_connection_table("DNa02", partners = "outputs")
dyak_connection_table("DNa02", partners = "outputs", summary = TRUE)
}
}
\seealso{
\code{\link[malevnc:manc_connection_table]{malevnc::manc_connection_table()}}
}
