% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{dyak_ids}
\alias{dyak_ids}
\title{Get yakuba ids in standard formats}
\usage{
dyak_ids(
  ids,
  mustWork = TRUE,
  as_character = TRUE,
  integer64 = FALSE,
  unique = FALSE,
  ...,
  dataset = dyak_default_dataset()
)
}
\arguments{
\item{ids}{Either numeric ids (in \code{character}, \code{numeric}, \code{integer} or
\code{integer64} format) or a query expression.}

\item{mustWork}{Whether to insist that at least one valid id is returned
(default \code{TRUE})}

\item{as_character}{Whether to return segments as character rather than
numeric vector (the default is character for safety).}

\item{integer64}{whether to return ids with class bit64::integer64.}

\item{unique}{Whether to ensure that only unique ids are returned (default
\code{TRUE})}

\item{...}{Additional arguments passed to \code{malevnc::manc_ids()}.}

\item{dataset}{The name of the dataset as reported in Clio or neuprint.}
}
\value{
A vector of numeric ids with mode determined by \code{as_character} and
\code{integer64}.
}
\description{
Get yakuba ids in standard formats
}
\examples{
\donttest{
ids <- dyak_ids("DNa02")
ids
dyak_ids("DNa02", integer64 = TRUE)
}
}
\seealso{
\code{\link[malevnc:manc_ids]{malevnc::manc_ids()}}
}
