% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodies.R
\name{dyak_islatest}
\alias{dyak_islatest}
\title{Check if a bodyid still exists in the specified yakuba DVID node}
\usage{
dyak_islatest(
  ids,
  node = "neutu",
  method = c("auto", "size", "sparsevol"),
  dataset = dyak_default_dataset()
)
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link[malevnc]{manc_ids}}}

\item{node}{A DVID node (defaults to the current neutu node, see
\code{\link[malevnc]{manc_dvid_node}})}

\item{method}{Which DVID endpoint to use. Expert use only.}

\item{dataset}{The name of the dataset as reported in Clio or neuprint.}
}
\value{
A logical vector ordered by input ids.
}
\description{
Check if a bodyid still exists in the specified yakuba DVID node
}
\details{
For details (and there are some) please see
\code{\link[malevnc:manc_islatest]{malevnc::manc_islatest()}}.
}
\examples{
\donttest{
dyak_islatest(c(10280,10490))
}
}
\seealso{
\code{\link[malevnc:manc_islatest]{malevnc::manc_islatest()}}
}
