% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{dyak_neuprint}
\alias{dyak_neuprint}
\title{Login to yakuba neuprint server}
\usage{
dyak_neuprint(
  token = Sys.getenv("neuprint_token"),
  dataset = NULL,
  Force = FALSE,
  ...
)
}
\arguments{
\item{token}{neuprint authorisation token obtained e.g. from
neuprint.janelia.org website.}

\item{dataset}{Allows you to override the neuprint dataset for
\code{dyak_neuprint()} (which would otherwise be chosen based on the value of
\code{options(yakuba.dataset)}, normally changed by \code{\link[=choose_dyak_dataset]{choose_dyak_dataset()}}).}

\item{Force}{Passed to \code{neuprintr::neuprint_login()}.}

\item{...}{Additional arguments passed to \code{\link{neuprint_login}}}
}
\value{
A \code{neuprint_connection} object returned by
\code{\link[neuprintr:neuprint_login]{neuprintr::neuprint_login()}}.
}
\description{
Login to yakuba neuprint server
}
\details{
It should be possible to use the same token across public and
private neuprint servers if you are using the same email address. However
this does not seem to work for all users. Before giving up on this, do try
using the \emph{most} recently issued token from a private server rather than
older tokens. If you need to pass a specific token you can use the \code{token}
argument, also setting \code{Force=TRUE} to ensure that the specified token is
used if you have already tried to log in during the current session.
}
\examples{
\donttest{
conn <- dyak_neuprint()
conn
}
}
\seealso{
\code{\link[malevnc:manc_neuprint]{malevnc::manc_neuprint()}}
}
