% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{dyak_neuprint_meta}
\alias{dyak_neuprint_meta}
\title{Fetch neuprint metadata for yakuba neurons}
\usage{
dyak_neuprint_meta(
  ids = NULL,
  conn = NULL,
  roiInfo = FALSE,
  simplify.xyz = TRUE,
  cache = NA,
  ...,
  dataset = dyak_default_dataset()
)
}
\arguments{
\item{ids}{body ids. When missing all bodies known to DVID are returned.}

\item{conn}{Optional, a \code{neuprint_connection} object. Defaults to
\code{\link[=dyak_neuprint]{dyak_neuprint()}} to ensure that query is against the yakuba dataset.}

\item{roiInfo}{whether to include the \code{roiInfo} field detailing synapse
numbers in different locations. This is omitted by default as it is
returned as a character vector of unprocessed JSON.}

\item{simplify.xyz}{Whether to simplify columns containing XYZ locations to a
simple \code{"x,y,z"} format rather than a longer form referencing a schema at
\code{spatialreference.org}. Defaults to \code{TRUE}.}

\item{cache}{whether to cache the query. When \code{cache=NA} (the default)
queries are cached for neuprint snapshot versions (but not production
datasets). See details.}

\item{...}{Additional arguments passed to \code{\link[neuprintr:neuprint_get_meta]{neuprintr::neuprint_get_meta()}}.}

\item{dataset}{The name of the dataset as reported in Clio or neuprint.}
}
\value{
A data frame with one row for each unique input id and \code{NA}s for all
columns except \code{bodyid} when neuprint holds no metadata.
}
\description{
Fetch neuprint metadata for yakuba neurons
}
\examples{
\donttest{
dyak_neuprint_meta("DNa02")
dyak_neuprint_meta(dyak_ids("DNa02"))
}
}
\seealso{
\code{\link[malevnc:manc_neuprint_meta]{malevnc::manc_neuprint_meta()}}
}
