% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{dyak_xyz2bodyid}
\alias{dyak_xyz2bodyid}
\title{Map XYZ locations to bodyids for the yakuba dataset}
\usage{
dyak_xyz2bodyid(
  xyz,
  units = c("raw", "nm", "microns", "um"),
  node = "neutu",
  cache = FALSE,
  dataset = dyak_default_dataset()
)
}
\arguments{
\item{xyz}{xyz location (by default in raw yakuba pixels)}

\item{units}{The optional units of the incoming 3D positions. Defaults to
\code{raw}.}

\item{node}{A DVID node as returned by \code{\link[malevnc]{manc_dvid_node}}. The
default is to return the current active (unlocked) node being used through
neutu.}

\item{cache}{Whether to cache the result of this call for 5 minutes.}

\item{dataset}{The name of the dataset as reported in Clio or neuprint.}
}
\value{
A character vector of body ids (\code{0} is missing somas / missing
locations).
}
\description{
Map XYZ locations to bodyids for the yakuba dataset
}
\examples{
\donttest{
meta <- dyak_neuprint_meta("DNa02")
dyak_xyz2bodyid(meta$somaLocation[1:2])
}
}
\seealso{
\code{\link[malevnc:manc_xyz2bodyid]{malevnc::manc_xyz2bodyid()}}
}
