% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodies.R
\name{read_dyak_neurons}
\alias{read_dyak_neurons}
\title{Read neuronal skeletons via neuprint}
\usage{
read_dyak_neurons(
  ids,
  connectors = FALSE,
  units = c("nm", "raw", "microns"),
  heal.threshold = Inf,
  conn = NULL,
  ...,
  dataset = dyak_default_dataset()
)
}
\arguments{
\item{ids}{bodyids in any form compatible with \code{malevnc::manc_ids()}.}

\item{connectors}{Whether to fetch synaptic connections for the neuron
(default \code{FALSE} in contrast to \code{\link[neuprintr:neuprint_read_neurons]{neuprintr::neuprint_read_neurons()}}).}

\item{units}{Units of the returned neurons (default \code{nm}).}

\item{heal.threshold}{The threshold for healing disconnected skeleton
fragments. The default of \code{Inf} ensures that all fragments are joined
together.}

\item{conn}{Optional, a \code{neuprint_connection} object. Defaults to
\code{\link[=dyak_neuprint]{dyak_neuprint()}} to ensure that query is against the yakuba dataset.}

\item{...}{Additional arguments passed to
\code{\link[neuprintr:neuprint_read_neurons]{neuprintr::neuprint_read_neurons()}}.}

\item{dataset}{The name of the dataset as reported in Clio or neuprint.}
}
\value{
A \code{nat::neuronlist} object containing one or more neurons.
}
\description{
Read neuronal skeletons via neuprint
}
\examples{
\donttest{
dns <- read_dyak_neurons("DNa02")
dns
}
}
\seealso{
\code{\link[malevnc:manc_read_neurons]{malevnc::manc_read_neurons()}}
}
