% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yakuba-package.R
\docType{package}
\name{yakuba-package}
\alias{yakuba}
\alias{yakuba-package}
\title{yakuba: Access the Drosophila yakuba FlyEM Dataset}
\description{
Provides a thin wrapper around the 'malevnc' package for working with the Drosophila yakuba male VNC neuprint and annotation datasets.
}
\section{Bridging registrations}{
 \code{yakuba} registers a bundled thin-plate
spline transform from \code{yakubaum} to \code{MANC} at package startup when
\code{nat.templatebrains} is available. A simple scaling alias is also added
from \code{yakuba} (nm) to \code{yakubaum} (microns), so
\code{nat.templatebrains::xform_brain()} can chain from \code{yakuba} to \code{MANC}.

Note the current registration (which lives in inst/extdata) was computed by
Sebastian Cachero in April 2025. See \href{https://flyem-cns.slack.com/archives/C07SFDC909W/p1744652271402679?thread_ts=1744645357.564199&cid=C07SFDC909W}{slack for details}

A second thin-plate spline registration from \code{yakubaum} to a symmetrised
yakuba template (\link{yakubasym}), also computed by Sebastian Cachero, supports
left-right mirroring via \code{\link[=mirror_dyak]{mirror_dyak()}} and \code{\link[=symmetric_dyak]{symmetric_dyak()}}.
}

\section{Package Options}{

There is currently one package option:

\itemize{
\item \code{yakuba.dataset} keeps track of the active yakuba dataset.
}

You can use this to run your code against the production yakuba dataset or
future snapshot variants. The main ways to do this are:

\enumerate{
\item Use \code{\link[=with_dyak]{with_dyak()}} to run one expression without changing the default
yakuba dataset.
\item Use \code{\link[=choose_dyak_dataset]{choose_dyak_dataset()}} to choose a default yakuba dataset for
the rest of the session.
\item Expert users may set the \code{yakuba.dataset} option directly in their
\code{.Rprofile}.
}
}

\examples{
\donttest{
options()[grepl("^yakuba", names(options()))]
}

\donttest{
library(malevnc)
library(nat)
library(nat.templatebrains)

MANC.tissue.surf.yak = nat.templatebrains::xform_brain(
  MANC.tissue.surf,
  sample = "MANC",
  reference = "yakuba"
)
wire3d(MANC.tissue.surf.yak)

dna02 = read_dyak_neurons("DNa02")
plot3d(dna02, lwd = 3)

# nb the default template space for MANC is calibrated in microns
# but for yakuba we use nm
dna02.manc = malevnc::manc_read_neurons("DNa02", units = "microns")
dna02.manc.yak = nat.templatebrains::xform_brain(
  dna02.manc,
  reference = "yakuba",
  sample = "MANC"
)
# set up nice coords for the plot
plot3d(dna02.manc.yak, col = "black", lwd = 2)
#' }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flyconnectome/yakuba}
  \item \url{https://flyconnectome.github.io/yakuba/}
  \item Report bugs at \url{https://github.com/flyconnectome/yakuba/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
}

}
\keyword{internal}
