% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{yakubasym}
\alias{yakubasym}
\title{Symmetric yakuba VNC template}
\format{
A \code{templatebrain} object (see \code{nat.templatebrains::templatebrain}).
}
\usage{
yakubasym
}
\description{
A \code{templatebrain} object describing the symmetrised yakuba VNC
template (in microns) used by \code{\link[=mirror_dyak]{mirror_dyak()}} and \code{\link[=symmetric_dyak]{symmetric_dyak()}}.
}
