% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{with_dyak}
\alias{with_dyak}
\alias{choose_dyak}
\title{Evaluate an expression after temporarily setting malevnc options}
\usage{
with_dyak(expr, dataset = dyak_default_dataset())

choose_dyak(dataset = dyak_default_dataset(), set = TRUE)
}
\arguments{
\item{expr}{An expression involving \code{malevnc}/\code{yakuba} functions to evaluate
with the specified autosegmentation.}

\item{dataset}{The name of the dataset as reported in Clio or neuprint.}

\item{set}{Whether to set the relevant package options or just to return a
list of the required options.}
}
\value{
\code{with_dyak()} returns the result of \code{expr}. \code{choose_dyak_dataset()}
and \code{choose_dyak()} return named lists of previous option values, matching
\code{options()}.
}
\description{
\code{yakuba} is a thin wrapper around \code{malevnc}. This function
temporarily changes the server/dataset options for \code{malevnc} while running
your expression.

\code{choose_dyak()} swaps out the default \code{malevnc} dataset for the
yakuba dataset. This means that all functions from the \code{malevnc} package
should target yakuba instead. It is recommended that you use \code{\link[=with_dyak]{with_dyak()}}
to do this temporarily unless you have no intention of using other FlyEM
datasets. \emph{To switch the default \code{yakuba} dataset please see
\code{choose_dyak_dataset()}.}
}
\details{
Note that it can also temporarily switch out the active dataset for
the \code{yakuba} package if you specify the \code{dataset} argument and this is
different from the current active dataset. This allows you to easily run
the same expression for different yakuba dataset variants.
}
\examples{
\donttest{
with_dyak(dyak_ids("DNa02"))

old <- choose_dyak("yakuba", set = FALSE)
str(old)
}
}
\seealso{
\code{\link[malevnc:choose_malevnc_dataset]{malevnc::choose_flyem_dataset()}},
\code{\link[malevnc:choose_malevnc_dataset]{malevnc::choose_malevnc_dataset()}}
}
