% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{yakuba_register_xforms}
\alias{yakuba_register_xforms}
\title{Register yakuba and MANC TPS transforms}
\usage{
yakuba_register_xforms()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
\code{yakuba_register_xforms()} registers the default transform
from \code{yakubaum} to \code{MANC} (see \code{\link[=xform_dyak2manc]{xform_dyak2manc()}}) and a simple scaling
alias from \code{yakuba} (nm) to \code{yakubaum} (microns) for use with
\code{nat.templatebrains::xform_brain()}. It also registers a thin-plate spline
from \code{yakubaum} to the symmetric template \code{yakubasym} (see
\code{\link[=mirror_dyak]{mirror_dyak()}}). Call once per session if the startup hook has not
already done so.
}
\examples{
\donttest{
library(malevnc)
library(nat)
library(nat.templatebrains)
yakuba_register_xforms()

MANC.tissue.surf.yak <- xform_brain(MANC.tissue.surf, sample = "MANC",
  reference = "yakuba")
wire3d(MANC.tissue.surf.yak)
}
}
