% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{yakuba_register_xforms}
\alias{yakuba_register_xforms}
\title{Register yakuba and MANC TPS transforms}
\usage{
yakuba_register_xforms()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
\code{yakuba_register_xforms()} registers a thin-plate spline
transform from \code{yakubaum} to \code{MANC} using the bundled
\code{yakuba_MANC_1000pts_tps.rds} file, and adds a simple scaling alias from
\code{yakuba} (nm) to \code{yakubaum} (microns). Call once per session if the startup
hook has not already done so.
}
\examples{
\donttest{
library(malevnc)
library(nat)
library(nat.templatebrains)
yakuba_register_xforms()

MANC.tissue.surf.yak <- xform_brain(MANC.tissue.surf, sample = "MANC",
  reference = "yakuba")
wire3d(MANC.tissue.surf.yak)
}
}
